Avoid following PURL redirects and split PURL tests into a separate file
Updated PURL-related tests to assert only the HTTP 302 status and Location header,
without following the redirect to the BioPortal UI. This avoids triggering
Cloudflare bot protection during CI runs.

We do not want to test UI behavior here — we only verify that the
correct redirect is served by the PURL resolver. Testing how the UI handles or
rewrites URLs (e.g., with query parameters or concept ID fragments) is out of scope
for these tests.

Split PURL resolution tests into a new file test_class_purl.rb to separate
concerns from core class model tests in test_class.rb

Resolves ncbo/ontologies_api_ruby_client/issues/41
